Introduce user defined MQTT keep alive value
`MQTT_KEEP_ALIVE_INTERVAL_SECONDS` macro value
depends on the responsiveness of the target
running the tests because some targets might
be faster or slower than others. Hence, it
would be useful to have a default value for
this macro (5) and the integration code can
provide other value if needed.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com